revert login dialog, move login disclaimer to a common component
